rename: ManualHandlerMapping 관련 클래스 패키지 이동 및 클래스 네이밍 변경
