+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -1,5 +1,6 @@
 package com.techcourse;
 
+import com.techcourse.mapping.ManualHandlerMappingWrapper;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
@@ -18,7 +19,7 @@ public class DispatcherServlet extends HttpServlet {
     private final transient HandlerMapping mapping;
 
     public DispatcherServlet() {
-        mapping = new ManualHandlerMappingAdapter();
+        mapping = new ManualHandlerMappingWrapper();
     }
 
     @Override
