+++ app/src/test/java/com/techcourse/ManualHandlerMappingTest.java
@@ -3,6 +3,7 @@ package com.techcourse;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import com.techcourse.mapping.ManualHandlerMapping;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
